feat(i18n): Add i18n configuration files
This commit adds the necessary configuration files for internationalization (i18n) support. It includes the i18n.config.ts file, which sets up the i18n options such as the fallback language and the path to the translation files. It also adds the i18n.constants.ts file, which defines constants used in the i18n configuration.

Signed-off-by: JohnARI <john.aristosa@hotmail.com>